+++ app/src/main/kotlin/com/kelsos/mbrc/networking/DiscoveryStop.kt
@@ -7,7 +7,7 @@ object DiscoveryStop {
   const val NOT_FOUND = 2
   const val COMPLETE = 3
 
-  @IntDef(NO_WIFI.toLong(), NOT_FOUND.toLong(), COMPLETE.toLong())
+  @IntDef(NO_WIFI, NOT_FOUND, COMPLETE)
   @Retention(AnnotationRetention.SOURCE)
   annotation class Reason
 }
